+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ttrpg-dice Changelog
 
+## [v0.3.0] - 2025-03-30
+
+### Breaking Changes
+
+- Directly subclass a specific `StatBlock` without using the `@statblock` decorator to create a specific set of stats as a class
+- `StatBlock` is no longer a `dataclass`
+
 ## [v0.2.0] - 2025-03-14
 
 ### Added